+++ shared/client/src/state/witness.rs
@@ -104,7 +104,7 @@ impl WitnessStep {
         let blooms = previous_round.blooms;
         let (participant_bloom, broadcast_bloom) = blooms.unwrap_or_default();
 
-        let mut client_times = FixedVec::<u64, { SOLANA_MAX_NUM_CLIENTS }>::default();
+        let mut client_times = FixedVec::<u16, { SOLANA_MAX_NUM_CLIENTS }>::default();
         for (_ident, time) in current_round.client_times.clone().into_iter() {
             let _ = client_times.push(time);
         }
